added crossword generator python app, moved js/css up a level out of the crossword_generator directory (which now hosts the app)
