+++ archive.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Manchat Crossword Archive</title>
-    <link rel="stylesheet" href="crossword-generator/crossword.css" />
+    <link rel="stylesheet" href="crossword.css" />
   </head>
   <body>
     <div class="archive-container">
@@ -33,6 +33,6 @@
       </div>
     </div>
 
-    <script src="crossword-generator/archive.js"></script>
+    <script src="archive.js"></script>
   </body>
 </html>
