--- crossword-generator/crossword.css
@@ -1,13 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Oswald:wght@400;500;600;700&family=Roboto+Condensed:wght@400;700&display=swap');
-@import url('css/base.css');
-@import url('css/layout.css');
-@import url('css/clues.css');
-@import url('css/feedback-toggle.css');
-@import url('css/leaderboard.css');
-@import url('css/mobile-clue-navigator.css');
-@import url('css/overlay.css');
-@import url('css/puzzle.css');
-@import url('css/timer.css');
-@import url('css/archive.css');
-@import url('css/statistics.css');
-@import url('css/puzzle-navigation.css');
