--- crossword-generator/css/clues.css
@@ -1,121 +1,0 @@
-.clues-section {
-  min-width: 350px;
-}
-
-.clues-section.blurred {
-  filter: blur(5px);
-  pointer-events: none;
-  user-select: none;
-  transition: filter 0.5s ease;
-}
-
-.clues-container {
-  display: flex;
-  flex-direction: row;
-  gap: 20px;    
-  flex-wrap: wrap;
-}
-
-.clue-group {
-  width: 47%;
-  min-width: 150px;
-}
-
-.clue-group h3 {
-  margin: 0 0 15px 0;
-  font-size: 24px;
-  color: var(--color-text-accent);
-  font-family: "Oswald", sans-serif;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  border-bottom: 3px solid var(--color-border-primary);
-  padding-bottom: 8px;
-  text-shadow: 1px 1px 2px var(--color-shadow-secondary);
-}
-
-.clue-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.clue-item {
-  padding: 6px 8px;
-  margin-bottom: 8px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  border-bottom: 1px solid var(--color-border-primary);
-  color: var(--color-text-primary);
-  font-weight: 500;
-}
-
-.clue-item:hover {
-  background: linear-gradient(
-    145deg,
-    var(--color-background-surface-hover),
-    var(--color-background-surface-hover-secondary)
-  );
-  border-color: var(--color-border-secondary);
-  box-shadow: 0 2px 8px var(--color-shadow-tertiary);
-}
-
-.clue-item.selected {
-  background: linear-gradient(
-    145deg,
-    var(--color-selected-background),
-    var(--color-selected-background-secondary)
-  );
-  color: var(--color-text-on-accent);
-  box-shadow: 0 4px 12px var(--color-shadow-ocean-light);
-}
-
-.clue-item.selected .clue-number {
-  color: var(--color-text-on-accent);
-}
-
-.clue-number {
-  font-weight: 700;
-  color: var(--color-text-accent);
-  font-family: "Oswald", sans-serif;
-}
-
-/* Tablet landscape - reduce cell size slightly */
-@media (max-width: 1024px) {
-  .clues-section {
-    min-width: 300px;
-  }
-}
-
-/* Tablet portrait - more aggressive scaling */
-@media (max-width: 768px) {
-  .clues-section {
-    min-width: 280px;
-  }
-}
-
-@media (min-width: 768px) {
-    .clues-section {
-        max-width: 50%;
-    }
-}
-
-/* Large mobile - force mobile layout */
-@media (max-width: 640px) {
-  .clues-section {
-    display: none; /* Hide desktop clues on mobile */
-  }
-}
-
-@media (max-width: 430px) {
-  .clues-section {
-    min-width: auto;
-    display: none; /* Hide desktop clues on mobile */
-  }
-}
-
-@media (min-width: 431px) and (max-width: 768px) and (max-height: 600px) {
-  .clues-section {
-    display: none; /* Hide desktop clues when keyboard is visible */
-  }
-}
