--- crossword-generator/css/puzzle.css
@@ -1,211 +1,0 @@
-.crossword-grid {
-  position: relative;
-  display: flex;
-  justify-content: center;
-}
-
-.grid {
-  display: grid;
-  gap: 2px;
-  max-width: 100%;
-  border-radius: 8px;
-}
-
-.cell-wrapper {
-  position: relative;
-  display: inline-block;
-}
-
-.cell {
-  width: var(--cell-size);
-  height: var(--cell-size);
-  background: linear-gradient(
-    145deg,
-    var(--color-background-cell),
-    var(--color-background-cell-secondary)
-  );
-  border: 0.5px solid var(--color-border-cell);
-  text-align: center;
-  font-size: var(--cell-font-size);
-  font-weight: 700;
-  font-family: "Franklin Gothic", sans-serif;
-  text-transform: uppercase;
-  position: relative;
-  cursor: pointer;
-  border-radius: 4px;
-  color: var(--color-text-cell);
-  transition: all 0.2s ease, width 0.3s ease, height 0.3s ease,
-    font-size 0.3s ease;
-  box-shadow: inset 0 1px 3px var(--color-shadow-light);
-  user-select: none;
-  -webkit-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  caret-color: transparent !important;
-}
-
-.cell::selection {
-  background: transparent;
-}
-
-.cell::-moz-selection {
-  background: transparent;
-}
-
-.cell.black {
-  background: linear-gradient(
-    145deg,
-    var(--color-background-cell-black),
-    var(--color-background-cell-black-secondary)
-  );
-  border: 2px solid var(--color-border-cell-black);
-  cursor: default;
-}
-
-.cell-wrapper.selected .cell {
-  outline: none;
-  background: linear-gradient(
-    145deg,
-    var(--color-background-cell-selected),
-    var(--color-background-cell-selected-secondary)
-  ) !important;
-  border: 6px solid var(--primitive-gold-600);
-  color: var(--color-text-cell) !important;
-  box-shadow: 0 0 0 2px var(--primitive-gold-400),
-    inset 0 1px 3px var(--color-shadow-light);
-  caret-color: transparent !important;
-}
-
-.cell.selected {
-  outline: none;
-  caret-color: transparent !important;
-}
-
-.cell.selected:focus {
-  outline: none;
-  caret-color: transparent !important;
-}
-
-.cell-wrapper.highlighted .cell {
-  background: linear-gradient(
-    145deg,
-    var(--color-highlighted-background-light),
-    var(--color-highlighted-background-light-secondary)
-  );
-  border: 0.5px solid var(--color-border-highlighted);
-  color: var(--color-text-selected-cell);
-  box-shadow: inset 0 1px 3px var(--color-shadow-light);
-}
-
-.cell-number {
-  position: absolute;
-  top: 6px;
-  left: 6px;
-  font-size: var(--cell-number-size);
-  font-weight: bold;
-  color: var(--color-text-cell-number);
-  pointer-events: none;
-  transition: font-size 0.3s ease;
-}
-
-/* Ensure selection is also transparent */
-input.cell::selection {
-  background: transparent !important;
-}
-
-input.cell::-moz-selection {
-  background: transparent !important;
-}
-
-/* Cell wrapper to contain input and cursor */
-.cell-wrapper {
-  width: var(--cell-size);
-  height: var(--cell-size);
-  position: relative;
-  border-radius: 4px;
-  transition: width 0.3s ease, height 0.3s ease;
-}
-
-.cell-wrapper .cell {
-  width: 100%;
-  height: 100%;
-}
-
-/* Mobile-specific text selection prevention and grid centering */
-@media (max-width: 768px) {
-  .crossword-grid {
-    margin: 0 auto;
-    width: 100%;
-    justify-content: center;
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .grid {
-    /* Ensure grid is properly centered within the available space */
-    margin: auto;
-    max-width: 100%;
-    max-height: 100%;
-  }
-
-  .cell {
-    -webkit-user-select: none !important;
-    -webkit-touch-callout: none !important;
-    -webkit-tap-highlight-color: transparent !important;
-    user-select: none !important;
-    -moz-user-select: none !important;
-    -ms-user-select: none !important;
-  }
-
-  .cell::selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  .cell::-moz-selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  /* Prevent highlighting on touch and focus */
-  .cell:focus::selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  .cell:focus::-moz-selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  /* Additional mobile-specific input handling */
-  input.cell {
-    -webkit-user-select: none !important;
-    -webkit-touch-callout: none !important;
-    -webkit-tap-highlight-color: transparent !important;
-    user-select: none !important;
-    -moz-user-select: none !important;
-    -ms-user-select: none !important;
-  }
-
-  input.cell::selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  input.cell::-moz-selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  input.cell:focus::selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-
-  input.cell:focus::-moz-selection {
-    background: transparent !important;
-    color: inherit !important;
-  }
-}
