--- crossword-generator/css/timer.css
@@ -1,304 +1,0 @@
-.timer-section {
-  text-align: center;
-  margin-bottom: 20px;
-  padding: 20px;
-  background: linear-gradient(
-    145deg,
-    var(--color-background-secondary),
-    var(--color-background-primary)
-  );
-  border-radius: 12px;
-  border: 2px solid var(--color-border-primary);
-  box-shadow: 0 4px 16px var(--color-shadow-secondary);
-}
-
-.timer {
-  font-size: 32px;
-  font-weight: 700;
-  font-family: "Oswald", sans-serif;
-  color: var(--color-text-accent);
-  margin-bottom: 15px;
-  text-shadow: 1px 1px 2px var(--color-shadow-secondary);
-  letter-spacing: 2px;
-}
-
-.timer-controls {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-/* Title area with mobile button */
-.title-area {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-}
-
-.mobile-options-btn {
-  display: none; /* Hidden by default, shown on mobile */
-  background: none;
-  border: none;
-  color: var(--color-text-accent);
-  font-size: 24px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  padding: 8px;
-  border-radius: 6px;
-  opacity: 0.8;
-}
-
-.mobile-options-btn:hover {
-  opacity: 1;
-  background: rgba(255, 255, 255, 0.1);
-  transform: scale(1.1);
-}
-
-.mobile-options-menu {
-  display: none;
-  position: relative;
-  z-index: 1000;
-  background: var(--color-background-overlay);
-  border: 2px solid var(--color-border-primary);
-  border-radius: 12px;
-  box-shadow: 0 8px 32px var(--color-shadow-primary);
-  backdrop-filter: blur(10px);
-  margin-top: 10px;
-  width: 100%;
-}
-
-.mobile-options-content {
-  padding: 15px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.mobile-timer-display {
-  text-align: center;
-  margin-bottom: 15px;
-  padding-bottom: 15px;
-  border-bottom: 2px solid var(--color-border-secondary);
-}
-
-.mobile-timer {
-  font-size: 28px;
-  font-weight: 700;
-  font-family: "Oswald", sans-serif;
-  color: var(--color-text-accent);
-  text-shadow: 1px 1px 2px var(--color-shadow-secondary);
-  letter-spacing: 2px;
-}
-
-.mobile-option-btn {
-  /* Match desktop .nav-btn styling exactly */
-  padding: 12px 24px;
-  font-size: 16px;
-  font-weight: 700;
-  font-family: "Oswald", sans-serif;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  border-radius: 6px;
-  border: 2px solid var(--color-border-primary);
-  background: linear-gradient(
-    145deg,
-    var(--color-background-surface),
-    var(--color-background-surface-secondary)
-  );
-  color: var(--color-text-accent);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  box-shadow: 0 4px 12px var(--color-shadow-tertiary);
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  width: 100%;
-}
-
-.mobile-option-btn:hover:not(:disabled) {
-  background: linear-gradient(
-    145deg,
-    var(--color-background-surface-hover),
-    var(--color-background-surface-hover-secondary)
-  );
-  border-color: var(--color-text-accent);
-  box-shadow: 0 6px 16px var(--color-shadow-secondary);
-  transform: translateY(-1px);
-}
-
-.mobile-option-btn:disabled {
-  background: linear-gradient(
-    145deg,
-    var(--color-disabled-background),
-    var(--color-disabled-background-secondary)
-  );
-  color: var(--color-disabled-text);
-  border-color: var(--color-disabled-border);
-  cursor: not-allowed;
-  transform: none;
-  box-shadow: none;
-}
-
-.mobile-feedback-toggle {
-  padding: 8px 0;
-  border-top: 1px solid var(--color-border-secondary);
-  margin-top: 5px;
-}
-
-.mobile-feedback-toggle label {
-  color: var(--color-text-secondary);
-  font-size: 14px;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  cursor: pointer;
-}
-
-.desktop-only {
-  display: block;
-}
-
-/* Mobile timer in header - displayed next to title on mobile only */
-.mobile-timer-header {
-  display: none; /* Hidden by default, shown on mobile */
-  font-size: 24px;
-  font-weight: 700;
-  font-family: "Oswald", sans-serif;
-  color: var(--color-text-accent);
-  text-shadow: 1px 1px 2px var(--color-shadow-secondary);
-  letter-spacing: 2px;
-  align-self: center;
-}
-
-/* Mobile layout - show mobile button and timer, hide desktop timer section */
-@media (max-width: 768px) {
-  .desktop-only {
-    display: none !important;
-  }
-
-  .mobile-options-btn {
-    display: block;
-  }
-
-  .mobile-timer-header {
-    display: block;
-  }
-
-  .title-area {
-    margin-bottom: 6px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-    gap: 8px;
-  }
-
-  .title {
-    flex: 1;
-    margin: 0;
-    text-align: left;
-  }
-}
-
-  .mobile-options-menu {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 90%;
-    max-width: 400px;
-    margin-top: 0;
-    z-index: 2000;
-  }
-  
-  /* Add overlay background when menu is open */
-  .mobile-options-menu::before {
-    content: '';
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: rgba(0, 0, 0, 0.5);
-    z-index: -1;
-  }
-
-
-/* Large mobile - adjust button size */
-@media (max-width: 640px) {
-  .mobile-options-btn {
-    font-size: 22px;
-    padding: 6px;
-  }
-}
-
-@media (max-width: 430px) {
-  .mobile-options-btn {
-    font-size: 20px;
-    padding: 4px;
-  }
-
-  .mobile-timer-header {
-    font-size: 20px;
-  }
-}
-
-@media (max-width: 430px) and (max-height: 600px) {
-  .timer-section {
-    padding: 6px;
-    margin-bottom: 6px;
-  }
-
-  .timer {
-    font-size: 20px;
-    margin-bottom: 4px;
-  }
-
-  .mobile-timer-header {
-    font-size: 18px;
-  }
-}
-
-@media (min-width: 431px) and (max-width: 768px) and (max-height: 600px) {
-  .timer-section {
-    padding: 8px;
-    margin-bottom: 8px;
-  }
-
-  .timer {
-    font-size: 22px;
-    margin-bottom: 6px;
-  }
-
-  .mobile-timer-header {
-    font-size: 20px;
-  }
-}
-
-@media (min-width: 431px) and (max-width: 768px) and (max-height: 400px) {
-  .timer-section {
-    padding: 8px;
-    margin-bottom: 8px;
-  }
-
-  .timer {
-    font-size: 22px;
-    margin-bottom: 6px;
-  }
-}
-
-@media (max-width: 320px) {
-  .timer {
-    font-size: 24px;
-  }
-
-  .timer-controls button {
-    padding: 8px 12px;
-    font-size: 12px;
-  }
-}
