+++ crossword.html
@@ -7,7 +7,7 @@
       content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no, viewport-fit=cover"
     />
     <title>Interactive Crossword</title>
-    <link rel="stylesheet" href="crossword-generator/crossword.css" />
+    <link rel="stylesheet" href="crossword.css" />
   </head>
   <body>
     <div class="page-container">
@@ -182,6 +182,6 @@
         </div>
       </div>
     </div>
-    <script src="crossword-generator/crossword.js"></script>
+    <script src="crossword.js"></script>
   </body>
 </html>
