+++ css/feedback-toggle.css
@@ -1,0 +1,54 @@
+.feedback-toggle {
+  margin-top: 10px;
+  text-align: center;
+}
+
+.feedback-toggle label {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 14px;
+  color: var(--color-text-muted);
+  cursor: pointer;
+}
+
+.feedback-toggle input[type='checkbox'] {
+  margin: 0;
+}
+
+/* Autocheck feedback styles with letterbox effect for selected/highlighted cells */
+
+/* Base autocheck feedback (when not selected/highlighted) */
+.cell.autocheck-correct {
+  background: #c8e6c9 !important;
+}
+
+.cell.autocheck-incorrect {
+  background: #ffcdd2 !important;
+}
+
+/* Letterbox styles for highlighted cells (selected word) with autocheck feedback */
+.cell-wrapper.highlighted .cell.autocheck-correct {
+  background: linear-gradient(145deg, #66bb6a, #4caf50) !important;
+  color: #ffffff !important;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+}
+
+.cell-wrapper.highlighted .cell.autocheck-incorrect {
+  background: linear-gradient(145deg, #ef5350, #f44336) !important;
+  color: #ffffff !important;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+}
+
+/* Letterbox styles for selected cell with autocheck feedback (darkest - most prominent) */
+.cell-wrapper.selected .cell.autocheck-correct {
+  background: linear-gradient(145deg, #388e3c, #2e7d32) !important;
+  color: #ffffff !important;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
+}
+
+.cell-wrapper.selected .cell.autocheck-incorrect {
+  background: linear-gradient(145deg, #d32f2f, #c62828) !important;
+  color: #ffffff !important;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
+}
