+++ css/layout.css
@@ -1,0 +1,127 @@
+.container {
+  display: flex;
+  gap: 30px;
+  background: linear-gradient(
+    145deg,
+    var(--color-background-surface),
+    var(--color-background-surface-secondary)
+  );
+  padding: 30px;
+  border-radius: 12px;
+  box-shadow:
+    0 8px 32px var(--color-shadow-secondary),
+    inset 0 1px 0 var(--color-shadow-inset-light);
+  border: 1px solid var(--color-border-primary);
+  justify-content: space-between;
+  margin: 0 auto;
+}
+
+.grid {
+  display: grid;
+  gap: 2px;
+  background-color: var(--color-background-grid);
+  border: 4px solid var(--color-background-primary);
+  width: fit-content;
+  max-width: 100%;
+  border-radius: 8px;
+  box-shadow: 0 4px 20px var(--color-shadow-primary);
+  margin: 0 auto;
+  /* Grid columns will be set dynamically via JavaScript */
+}
+
+/* Mobile crossword grid - flexible sizing and hide clues */
+@media (max-width: 768px) {
+  .crossword-grid {
+    flex: 1; /* Take up all available space */
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    min-height: 0; /* Allow shrinking */
+    margin-bottom: 0; /* No bottom margin needed */
+  }
+
+  /* Hide the desktop clues section on mobile */
+  .clues-section {
+    display: none;
+  }
+}
+
+/* Tablet landscape - adjust layout */
+@media (max-width: 1024px) {
+  .clues-section {
+    min-width: 300px;
+  }
+
+  .container {
+    padding: 20px;
+    gap: 20px;
+  }
+}
+
+/* Tablet portrait - more aggressive scaling */
+@media (max-width: 768px) {
+  .container {
+    gap: 15px;
+    padding: 15px;
+    flex-direction: column;
+    gap: 20px;
+  }
+}
+
+/* Large mobile - force mobile layout */
+@media (max-width: 640px) {
+  .container {
+    flex-direction: column; /* Force vertical stacking on all mobile devices */
+    gap: 8px; /* Reduced gap for more compact layout */
+    padding: 6px; /* Reduced padding */
+    margin: 0; /* Remove bottom margin - no scrolling needed */
+    max-width: 100%;
+    flex: 1; /* Take up remaining space in page-container */
+    overflow: hidden; /* Prevent any overflow scrolling */
+    min-height: 0; /* Allow flex shrinking */
+    padding-bottom: 80px; /* Account for fixed bottom clue navigator */
+  }
+}
+
+@media (max-width: 430px) {
+  .container {
+    flex-direction: column;
+    gap: 6px; /* Even more compact for small screens */
+    padding: 4px; /* Minimal padding */
+    margin: 0; /* Remove bottom margin - no scrolling needed */
+    max-width: 100%;
+    flex: 1; /* Take up remaining space in page-container */
+    overflow: hidden; /* Prevent any overflow scrolling */
+    min-height: 0; /* Allow flex shrinking */
+    padding-bottom: 80px; /* Account for fixed bottom clue navigator */
+  }
+}
+
+@media (max-width: 430px) and (max-height: 600px) {
+  .container {
+    gap: 6px;
+    padding: 3px;
+    margin: 0 0 400px 0; /* Extra space when keyboard is likely visible */
+    /* Ensure container allows mobile clue navigator to be visible */
+    position: relative;
+    overflow: visible;
+  }
+}
+
+@media (min-width: 431px) and (max-width: 768px) and (max-height: 600px) {
+  .container {
+    gap: 8px;
+    padding: 5px;
+    margin: 0 0 200px 0;
+    flex-direction: column;
+    overflow: visible;
+    min-height: 360px;
+  }
+}
+
+@media (max-width: 320px) {
+  .container {
+    padding: 8px;
+  }
+}
