+++ css/overlay.css
@@ -1,0 +1,92 @@
+.game-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(
+    135deg,
+    var(--color-background-overlay),
+    var(--color-background-overlay-secondary)
+  );
+  backdrop-filter: blur(15px);
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  z-index: 2000;
+  animation: fadeIn 0.5s ease;
+}
+
+.overlay-content {
+  background: linear-gradient(
+    145deg,
+    var(--color-background-surface),
+    var(--color-background-surface-secondary)
+  );
+  padding: 50px;
+  border-radius: 16px;
+  box-shadow:
+    0 24px 48px var(--color-shadow-primary),
+    inset 0 1px 0 var(--color-shadow-inset-light);
+  border: 2px solid var(--color-border-primary);
+  text-align: center;
+  max-width: 600px;
+  animation: slideUp 0.6s ease;
+}
+
+.overlay-content h2 {
+  margin: 0 0 25px 0;
+  font-size: 36px;
+  line-height: 36px;
+  color: var(--color-text-accent);
+  font-family: 'Oswald', sans-serif;
+  font-weight: 700;
+  text-transform: uppercase;
+  letter-spacing: 3px;
+  text-shadow: 2px 2px 4px var(--color-shadow-secondary);
+  text-align: center;
+}
+
+.overlay-content p {
+  margin: 0 0 35px 0;
+  font-size: 18px;
+  color: var(--color-text-secondary);
+  line-height: 1.6;
+  font-family: 'Roboto Condensed', sans-serif;
+  font-weight: 400;
+}
+
+.start-game-btn {
+  padding: 18px 40px;
+  font-size: 20px;
+  font-weight: 700;
+  font-family: 'Oswald', sans-serif;
+  text-transform: uppercase;
+  letter-spacing: 2px;
+  background: linear-gradient(
+    145deg,
+    var(--color-selected-background),
+    var(--color-selected-background-secondary)
+  );
+  color: var(--color-text-on-accent);
+  border-radius: 12px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+}
+
+.start-game-btn:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 8px 24px var(--color-shadow-ocean-hover);
+  background: linear-gradient(
+    145deg,
+    var(--color-hover-background),
+    var(--color-hover-background-secondary)
+  );
+}
+
+/* Large Tablet and above */
+@media (min-width: 768px) {
+  .game-overlay {
+    align-items: center;
+  }
+}
