+++ statistics.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Manchat Crossword Statistics</title>
-    <link rel="stylesheet" href="crossword-generator/crossword.css" />
+    <link rel="stylesheet" href="crossword.css" />
   </head>
   <body>
     <div class="archive-container">
@@ -58,6 +58,6 @@
       </div>
     </div>
 
-    <script src="crossword-generator/statistics.js"></script>
+    <script src="statistics.js"></script>
   </body>
 </html>
